+++ src/adplugins/adtextfile/txt_tokenizer.hpp
@@ -37,7 +37,7 @@ namespace adtextfile {
         typedef std::tuple< std::vector< double >    // time
                             , std::vector< double >  // mass
                             , std::vector< double >  // intensity
-                            , std::vector< int >     // color
+                            , std::vector< uint8_t > // color
                             > data_type;
         typedef std::array< bool, 4 > flags_type;
